fix:  lms installed apps without course_creators
Due course_creators is an studio app, when lms use `eox_nelp` it bring an error.
This avoid the following error ```RuntimeError: Model class cms.djangoapps.course_creators.models.CourseCreator doesn't declare an explicit app_label and isn't in an application in INSTALLED_APPS```.
So you have to add this one in the `INSTALLED_APPS`.