feat: positive logic to the stats view
By default the view show all components. Then you can manage or filter which component
wouldnt be show setting it to false.

feat: use frontend as experience use it

This would be considered in future of the need of a refactor,
but by the moment is better to have the templates views that use react
with the same shape.

chore: apply suggestions from code review